+++ packages/juice_routing/example/lib/auth_bloc.dart
@@ -70,10 +70,12 @@ class AuthBloc extends JuiceBloc<AuthState> {
             () => UseCaseBuilder(
                   typeOfEvent: LoginEvent,
                   useCaseGenerator: () => LoginUseCase(),
+                  concurrency: EventConcurrency.droppable,
                 ),
             () => UseCaseBuilder(
                   typeOfEvent: LogoutEvent,
                   useCaseGenerator: () => LogoutUseCase(),
+                  concurrency: EventConcurrency.sequential,
                 ),
           ],
         );